Remove Flux version of blocks/follow-button
